Move `fetch_sharded` into `fetch`
