+++ src/fetch.rs
@@ -2,14 +2,15 @@
 // rather than the current which is not limited and may cause bottlenecks
 use anyhow::Context;
 use byteorder::{ByteOrder, LittleEndian};
+use futures::future::join_all;
 use reqwest::{
     header::{HeaderValue, RANGE},
     Client, StatusCode,
 };
-use std::collections::HashMap;
+use std::collections::{HashMap, HashSet};
 
-use crate::errors::HubError;
-use crate::schemas::FileType;
+use crate::errors::RequestError;
+use crate::schemas::{FileType, ModelIndex};
 
 // As per https://github.com/huggingface/huggingface_hub/pull/1855#discussion_r1404286419, we
 // only fetch the first 100kb of the `model.safetensors` file, as empirically, 97% of
@@ -21,7 +22,7 @@ async fn fetch_metadata(
     url: &str,
     range_lower: Option<usize>,
     range_upper: Option<usize>,
-) -> anyhow::Result<(bytes::Bytes, usize), HubError> {
+) -> anyhow::Result<(bytes::Bytes, usize), RequestError> {
     let range_lower = range_lower.unwrap_or(0);
     let range_upper = range_upper.unwrap_or(MAX_METADATA_SIZE);
 
@@ -36,19 +37,25 @@ async fn fetch_metadata(
         .await
         .context("fetching the provided url failed")
     {
-        Ok(response) => match response.status() {
-            StatusCode::PARTIAL_CONTENT => {
-                let metadata = response
-                    .bytes()
-                    .await
-                    .context("failed reading the bytes from the response")?;
-                let metadata_size = LittleEndian::read_u64(&metadata[..8]) as usize;
-                Ok((metadata, metadata_size))
+        Ok(response) => {
+            let status_code = response.status();
+            match status_code {
+                StatusCode::PARTIAL_CONTENT => {
+                    let metadata = response
+                        .bytes()
+                        .await
+                        .context("failed reading the bytes from the response")?;
+                    let metadata_size = LittleEndian::read_u64(&metadata[..8]) as usize;
+                    Ok((metadata, metadata_size))
+                }
+                StatusCode::NOT_FOUND => Err(RequestError::FileNotFound(url.to_string())),
+                StatusCode::FORBIDDEN => Err(RequestError::HubAuth),
+                StatusCode::INTERNAL_SERVER_ERROR => Err(RequestError::Internal),
+                StatusCode::SERVICE_UNAVAILABLE => Err(RequestError::HubIsDown),
+                _ => Err(RequestError::Unknown(status_code)),
             }
-            StatusCode::NOT_FOUND => Err(HubError::FileNotFound(url.to_string())),
-            _ => Err(HubError::Internal),
-        },
-        Err(e) => Err(HubError::Other(e)),
+        }
+        Err(e) => Err(RequestError::Other(e)),
     }
 }
 
@@ -57,7 +64,7 @@ pub async fn fetch(
     model_id: String,
     revision: Option<String>,
     filename: Option<String>,
-) -> anyhow::Result<HashMap<String, FileType>, HubError> {
+) -> anyhow::Result<HashMap<String, FileType>> {
     let url = format!(
         "https://huggingface.co/{}/resolve/{}/{}",
         model_id,
@@ -82,4 +89,82 @@ pub async fn fetch(
         serde_json::from_slice::<HashMap<String, FileType>>(&metadata[8..metadata_size + 8])
             .context("parsing the metadata bytes into a serde_json::Value failed")?,
     )
+}
+
+pub async fn fetch_sharded(
+    client: &Client,
+    model_id: String,
+    revision: Option<String>,
+) -> anyhow::Result<HashMap<String, FileType>, RequestError> {
+    let url = format!(
+        "https://huggingface.co/{}/resolve/{}/model.safetensors.index.json",
+        model_id,
+        revision.clone().unwrap_or("main".to_string()),
+    );
+
+    match client
+        .get(&url)
+        .send()
+        .await
+        .context("fetching the sharded url failed")
+    {
+        Ok(response) => {
+            let status_code = response.status();
+            match status_code {
+                StatusCode::OK => {
+                    let model_index = serde_json::from_str::<ModelIndex>(
+                        &response
+                            .text()
+                            .await
+                            .context("failed to pull the text out of the response")?,
+                    )
+                    .context("failed deserializing the text into a json")?;
+
+                    let filenames = model_index
+                        .weight_map
+                        .values()
+                        .cloned()
+                        .collect::<HashSet<String>>();
+
+                    let mut tasks = Vec::new();
+                    for filename in filenames {
+                        let client_clone = client.clone();
+                        let model_id_clone = model_id.clone();
+                        let revision_clone = revision.clone();
+                        tasks.push(tokio::spawn(async move {
+                            fetch(
+                                &client_clone,
+                                model_id_clone,
+                                revision_clone,
+                                Some(filename),
+                            )
+                            .await
+                            .context(
+                                "failed to fetch metadata from the safetensors file from the hub",
+                            )
+                        }));
+                    }
+
+                    let mut metadata = HashMap::<String, FileType>::new();
+                    let futures = join_all(tasks).await;
+                    for future in futures {
+                        match future {
+                            Ok(Ok(result)) => {
+                                metadata.extend(result);
+                            }
+                            Ok(Err(..)) => panic!("failed capturing future"),
+                            Err(..) => panic!("failed capturing future"),
+                        }
+                    }
+                    Ok(metadata)
+                }
+                StatusCode::NOT_FOUND => Err(RequestError::FileNotFound(url.to_string())),
+                StatusCode::FORBIDDEN => Err(RequestError::HubAuth),
+                StatusCode::INTERNAL_SERVER_ERROR => Err(RequestError::Internal),
+                StatusCode::SERVICE_UNAVAILABLE => Err(RequestError::HubIsDown),
+                _ => Err(RequestError::Unknown(status_code)),
+            }
+        }
+        Err(e) => Err(RequestError::Other(e)),
+    }
 }
